Fix duplicate command name.
